# specs/005-swagger-api-docs/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: "API-X"
  version: "0.1.0"
  description: |
    **API-X** adalah backend REST API untuk manajemen event dan pendaftaran peserta,
    dilengkapi dengan autentikasi berbasis JWT dan laporan statistik untuk admin.

    ## Cara Menggunakan API Ini

    Sebagian besar endpoint memerlukan autentikasi. Ikuti langkah berikut:

    1. **Daftar akun** — Panggil `POST /auth/register` dengan nama, email, dan kata sandi.
    2. **Login** — Panggil `POST /auth/login` dengan email dan kata sandi untuk mendapatkan JWT access token.
    3. **Salin token** — Dari respons login, salin nilai `access_token`.
    4. **Masukkan token** — Klik tombol **Authorize** di atas, lalu masukkan token dalam format `Bearer <token>`.
    5. **Akses endpoint** — Semua endpoint terproteksi kini dapat digunakan langsung dari halaman ini.

    ## Peran Pengguna

    | Peran | Keterangan |
    |-------|-----------|
    | **Publik** | Dapat diakses tanpa autentikasi |
    | **Memerlukan token** | Memerlukan JWT token pengguna terdaftar |
    | **Khusus admin** | Memerlukan JWT token dengan role `admin` |

    ## Format Error

    Semua error menggunakan struktur JSON yang konsisten:

    ```json
    {
      "error": {
        "code": "KODE_ERROR",
        "message": "Pesan yang dapat dibaca manusia.",
        "httpStatus": 4xx
      }
    }
    ```
servers:
  - url: "http://localhost:8000"
    description: "Lokal (development)"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token yang diperoleh dari `POST /auth/login`.

        **Cara mendapatkan token**:
        1. Panggil `POST /auth/login` dengan email dan kata sandi yang valid.
        2. Salin nilai `access_token` dari respons.
        3. Klik tombol **Authorize** di atas dan masukkan nilai token (tanpa awalan "Bearer").

        **Format header**: `Authorization: Bearer <token>`

        Token berlaku selama **60 menit** sejak diterbitkan (HS256 JWT).
  schemas:
...
    RegisterRequest:
      type: object
      required: [name, email, password]
      description: "Data yang diperlukan untuk mendaftarkan akun pengguna baru."
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: "Nama lengkap pengguna."
          examples:
            - "Budi Santoso"
        email:
          type: string
          format: email
          description: "Alamat email unik yang digunakan untuk login. Tidak boleh sama dengan email yang sudah terdaftar."
          examples:
            - "budi@contoh.com"
        password:
          type: string
          minLength: 8
          maxLength: 72
          description: "Kata sandi (minimal 8 karakter, maksimal 72 karakter). Disimpan sebagai hash bcrypt — tidak pernah disimpan dalam bentuk plain text."
          examples:
            - "R4h4sia!Kuat"
    LoginRequest:
      type: object
      required: [email, password]
      description: "Kredensial untuk login dan mendapatkan JWT access token."
      properties:
        email:
          type: string
          format: email
          description: "Alamat email akun yang sudah terdaftar."
          examples:
            - "budi@contoh.com"
        password:
          type: string
          minLength: 1
          description: "Kata sandi akun."
          examples:
            - "R4h4sia!Kuat"
    TokenResponse:
      type: object
      description: "JWT access token yang dikembalikan setelah login berhasil."
      properties:
        access_token:
          type: string
          description: "JWT access token. Sertakan di header `Authorization: Bearer <token>` untuk mengakses endpoint terproteksi. Berlaku selama 60 menit sejak diterbitkan."
          examples:
            - "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJmNDdhYzEwYi01OGNjLTQzNzItYTU2Ny0wZTAyYjJjM2Q0NzkiLCJyb2xlIjoidXNlciIsImlhdCI6MTc0MTUwMDAwMCwiZXhwIjoxNzQxNTAzNjAwfQ.signature"
        token_type:
          type: string
          description: "Tipe token. Selalu `\"bearer\"`."
          examples:
            - "bearer"
    UserRead:
      type: object
      description: "Data profil pengguna yang dikembalikan oleh sistem."
      properties:
        id:
          type: string
          format: uuid
          description: "ID unik pengguna (UUID)."
          examples:
            - "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        name:
          type: string
          description: "Nama lengkap pengguna."
          examples:
            - "Budi Santoso"
        email:
          type: string
          format: email
          description: "Alamat email pengguna."
          examples:
            - "budi@contoh.com"
        role:
          type: string
          enum: [user, admin]
          description: "Peran pengguna. `\"user\"` untuk pengguna biasa; `\"admin\"` untuk administrator."
          examples:
            - "user"
        created_at:
          type: string
          format: date-time
          description: "Waktu pembuatan akun (UTC, format ISO 8601)."
          examples:
            - "2026-03-01T08:00:00Z"

    # ------------------------------------------------------------------ Events
...
    ErrorDetail:
      type: object
      description: "Detail error yang dikembalikan oleh API."
      required: [code, message, httpStatus]
      properties:
        code:
          type: string
          description: "Kode error yang unik dan dapat dibaca program. Gunakan nilai ini untuk penanganan error di sisi klien."
          examples:
            - "UNAUTHORIZED"
        message:
          type: string
          description: "Pesan error yang dapat dibaca manusia, menjelaskan apa yang terjadi."
          examples:
            - "Could not validate credentials."
        httpStatus:
          type: integer
          description: "Kode HTTP status, sama dengan status code respons."
          examples:
            - 401
    ErrorEnvelope:
      type: object
      description: "Struktur JSON standar untuk semua respons error di API-X."
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
      examples:
        unauthorized:
          summary: "401 Unauthorized"
          value:
            error:
              code: "UNAUTHORIZED"
              message: "Could not validate credentials."
              httpStatus: 401
        forbidden:
          summary: "403 Forbidden"
          value:
            error:
              code: "FORBIDDEN"
              message: "You do not have permission to perform this action."
              httpStatus: 403
        not_found:
          summary: "404 Not Found"
          value:
            error:
              code: "EVENT_NOT_FOUND"
              message: "Event not found."
              httpStatus: 404
        validation_error:
          summary: "422 Validation Error"
          value:
            error:
              code: "VALIDATION_ERROR"
              message: "value is not a valid email address"
              httpStatus: 422
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: "Daftarkan akun pengguna baru"
      description: |
        Membuat akun pengguna baru dengan nama, email, dan kata sandi.

        **Akses**: Publik — tidak memerlukan autentikasi.

        **Aturan**:
        - Email harus unik; tidak boleh sama dengan email yang sudah terdaftar.
        - Kata sandi minimal 8 karakter dan disimpan sebagai hash bcrypt (tidak pernah plain text).
        - Akun baru secara default memiliki role `user`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              name: "Budi Santoso"
              email: "budi@contoh.com"
              password: "R4h4sia!Kuat"
      responses:
        "201":
          description: "Akun berhasil dibuat. Mengembalikan data profil pengguna baru."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRead"
              example:
                id: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                name: "Budi Santoso"
                email: "budi@contoh.com"
                role: "user"
                created_at: "2026-03-09T08:00:00Z"
        "409":
          description: "Email sudah digunakan oleh akun lain."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "EMAIL_ALREADY_EXISTS"
                  message: "An account with this email already exists."
                  httpStatus: 409
        "422":
          description: "Data request tidak valid (format email salah, kata sandi terlalu pendek, dll.)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "VALIDATION_ERROR"
                  message: "value is not a valid email address"
                  httpStatus: 422
  /auth/login:
    post:
      tags: [auth]
      summary: "Login dan dapatkan JWT access token"
      description: |
        Memverifikasi email dan kata sandi, lalu mengembalikan JWT access token
        yang digunakan untuk mengakses endpoint terproteksi.

        **Akses**: Publik — tidak memerlukan autentikasi.

        **Penting**: Setelah mendapatkan token, gunakan tombol **Authorize** di atas
        atau sertakan di header: `Authorization: Bearer <token>`.

        Token berlaku selama **60 menit** sejak diterbitkan.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: "budi@contoh.com"
              password: "R4h4sia!Kuat"
      responses:
        "200":
          description: "Login berhasil. Mengembalikan JWT access token."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
              example:
                access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJmNDdhYzEwYi01OGNjLTQzNzItYTU2Ny0wZTAyYjJjM2Q0NzkiLCJyb2xlIjoidXNlciIsImlhdCI6MTc0MTUwMDAwMCwiZXhwIjoxNzQxNTAzNjAwfQ.signature"
                token_type: "bearer"
        "401":
          description: "Kredensial tidak valid (email tidak terdaftar atau kata sandi salah)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "UNAUTHORIZED"
                  message: "Could not validate credentials."
                  httpStatus: 401
        "422":
          description: "Data request tidak valid."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "VALIDATION_ERROR"
                  message: "value is not a valid email address"
                  httpStatus: 422
  /auth/me:
    get:
      tags: [auth]
      summary: "Lihat profil pengguna yang sedang login"
      description: |
        Mengembalikan data profil pengguna yang saat ini terautentikasi berdasarkan JWT token.

        **Akses**: Memerlukan token — sertakan JWT di header `Authorization: Bearer <token>`.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: "Profil pengguna berhasil diambil."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRead"
              example:
                id: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                name: "Budi Santoso"
                email: "budi@contoh.com"
                role: "user"
                created_at: "2026-03-01T08:00:00Z"
        "401":
          description: "Token tidak ada, tidak valid, atau sudah kedaluwarsa."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "UNAUTHORIZED"
                  message: "Could not validate credentials."
                  httpStatus: 401

  # ================================================================ Admin Users
...